+++ src/asset_definitions/media.py
@@ -7,6 +7,7 @@ from typing import *  # noqa
 
 __all__ = (
     "Media",
+    "MediaDefiningClass",
 )
 
 
@@ -57,3 +58,20 @@ class Media(django.forms.Media):
         media = Media(js=self._media._js, css=self._media._css)
         media._combined_with = self._combined_with + [other]
         return media
+
+
+class MediaDefiningClass(object):
+
+    @property
+    def media(self):
+        return self.get_media()
+
+    def get_media(self):
+        # type: () -> Media
+        return self._get_media_from_definition() or Media()
+
+    def _get_media_from_definition(self):
+        # type: () -> Optional[Media]
+        definition = getattr(self.__class__, "Media", None)
+        if definition:
+            return Media(media=definition)